Add dedication to book.tex and pdf
Solves this issue for the compiled pdf: https://github.com/gvwilson/teachtogether.tech/issues/274

I also added a line to the MakeFile so that _book.pdf_ is copied to _libro_enproceso.pdf_. 
book.* is listed in gitignore, but not libro_enproceso. I find useful to access to the pdf version through  github.